+++ src/main/java/ck4/nvb/rsmanagement/core/module/stores/product/service/dto/ProductGetDto.java
@@ -9,7 +9,7 @@ public class ProductGetDto extends BaseProductDto {
     private String sku;
     private String name;
     private String description;
-    private double unitPrice;
+    private Integer unitPrice;
     private Long categoryId;
     private Long supplierId;
 }
